requirements.txt
pyaudio
SpeechRecognition
python-dotenv
ollama
openai
pygame
RealtimeSTT
python-weather
icrawler
spotipy
--extra-index-url https://download.pytorch.org/whl/cu116
torch
git+https://github.com/openai/whisper.git
pyannote.audio
numpy
scipy
soundfile
librosa
scikit-learn
# CSM TTS dependencies
transformers>=4.53.0
huggingface_hub
accelerate
datasets
# Kyutai STT dependencies (requires transformers >= 4.53.0)
# Note: The Kyutai model requires the latest transformers version